fix: Tooltip컴포넌트가 내용을 props로 받아오도록 변경
